% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acthd_get_color.R
\name{acthd_get_color}
\alias{acthd_get_color}
\title{Function return ACTHD color}
\usage{
acthd_get_color(color = "bgs blue")
}
\arguments{
\item{color}{Name of the colour to return. See \code{\link[=acthd_list_colors]{acthd_list_colors()}} for
available names.}
}
\description{
Function return ACTHD color
}
