% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acthd_ggplot_autotheme.R
\name{acthd_ggplot_autotheme}
\alias{acthd_ggplot_autotheme}
\title{Auto format ggplot object}
\usage{
acthd_ggplot_autotheme(x)
}
\arguments{
\item{x}{\code{gt()} object to have theme applied to it.}
}
\description{
This function allows the user to apply pre-defined theme to a ggplot object.
}
\examples{
# Two-step application
output = ggplot2::mpg |> 
  ggplot2::ggplot(ggplot2::aes(x = manufacturer, y=hwy)) +
  ggplot2::geom_col()
output |> 
  acthd_ggplot_autotheme()
  
# Single-step application
{ output = ggplot2::mpg |> 
    ggplot2::ggplot(ggplot2::aes(x = manufacturer, y=hwy)) +
    ggplot2::geom_col()
} |> 
  acthd_ggplot_autotheme()
 
}
\keyword{acthd}
\keyword{gpplot2}
