% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acthd_ggplot_theme.R
\name{acthd_ggplot_theme}
\alias{acthd_ggplot_theme}
\title{Color scale constructor for ACT Health Directorate palettes}
\usage{
acthd_ggplot_theme(flip = FALSE)
}
\arguments{
\item{flip}{Boolean to indicate if the ggplot object has had \code{coord_flip()} applied to it.}
}
\description{
This function allows the user to apply pre-defined theme to a ggplot object.
}
\examples{
# Normal theme application
ggplot2::ggplot(ggplot2::mpg, ggplot2::aes(manufacturer, fill = manufacturer)) +
  ggplot2::geom_bar() +
  ggplot2::coord_flip() +
  actepir:::acthd_ggplot_theme(flip=TRUE)
  
# Flipped application
ggplot2::ggplot(ggplot2::mpg, ggplot2::aes(manufacturer, fill = manufacturer)) +
  ggplot2::geom_bar() +
  ggplot2::coord_flip() +
  actepir:::acthd_ggplot_theme(flip=TRUE)
 
}
\keyword{gpplot2}
\keyword{internal}
