% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acthd_list_colors.R
\name{acthd_list_colors}
\alias{acthd_list_colors}
\title{Get ACTHD colour details}
\usage{
acthd_list_colors(detail = "names", objtype = "list")
}
\arguments{
\item{detail}{What level of detail to be returned.  Valid options are \code{names} or \code{all}. Default is \code{names}.}

\item{objtype}{The type of object to be returned.  Valid options are \code{list} or \code{df}.  Default is \code{list}.}
}
\description{
This function will allow the user to get colour details.
}
\examples{
# get a simple list of color names
acthd_list_colors()

}
\keyword{colors}
