% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acthd_palettes.R
\name{acthd_palettes}
\alias{acthd_palettes}
\title{Function to store/return ACTHD palettes}
\usage{
acthd_palettes(palette = "blue_turq", reverse = FALSE, ...)
}
\description{
Function to store/return ACTHD palettes
}
