% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autofactor.R
\name{autofactor}
\alias{autofactor}
\title{Automatic factor conversion}
\usage{
autofactor(
  dfdata = parent.frame(),
  dflabels,
  vars,
  order = NULL,
  suffix = NULL,
  nolabel = NULL
)
}
\arguments{
\item{dfdata}{The dataframe containing the column(s) requiring factor conversion.}

\item{dflabels}{The dataframe containing the data dictionary/labels.}

\item{vars}{A vector containing the column names that should be factor converted. Argument should be supplied as character.}

\item{order}{Should the resulting factor levels be ordered?  Argument should be supplied as TRUE/FALSE.}

\item{suffix}{The suffix to the resulting factor column name to distinguish it from the original. Argument should be supplied as character. Not supplying will overwrite the original column.}

\item{nolabel}{The name of the level which unmatched levels are assigned to.  Argument should be supplied as character.}
}
\description{
This function will allow the user to convert a column to a factor and automatically set levels based on a supplied data dictionary.
}
\details{
The dictionary must conform to the DataLabels structure used on EpiServer.
}
\examples{
# Labels in the EpiServer DataLabels structure
mylabelsdf <- data.frame(
  VarName   = c("var1", "var1", "var2", "var2"),
  LabelType = "OPT",
  DataType  = "INT",
  DataCode  = c(1, 2, 1, 2),
  LabelName = c("Male", "Female", "No", "Yes")
  )
# The value 3 in var2 has no label and is assigned to `nolabel`
mydata <- data.frame(var1 = c(1, 2, 1), var2 = c(2, 3, 1))

varlist <- c("var1","var2")
mydata <- mydata |> 
  autofactor(
    dflabels = mylabelsdf,
    vars = varlist,
    order = TRUE,
    suffix = "_X",
    nolabel = "UL"
    )
    
}
\keyword{autofactor}
