% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_withlabels.R
\name{collect_withlabels}
\alias{collect_withlabels}
\title{Collect in dbplyr pipe with labels}
\usage{
collect_withlabels(
  dataset,
  lbl_db = NULL,
  lbl_schema = NULL,
  lbl_table = NULL,
  driver = NULL,
  max_attempts = NULL,
  ...
)
}
\arguments{
\item{dataset}{The name of the dataset that labels are associated with. Argument
should be supplied as character or object.}

\item{lbl_db}{The name of the database to be connected to. Argument should be
supplied as character.}

\item{lbl_schema}{The name of the schema to be connected to. Argument should be
supplied as character.}

\item{lbl_table}{The name of the table where the labels exist. Argument should
be supplied as character.}

\item{driver}{Character string or \code{NULL}. ODBC driver to use for the connection.
If \code{NULL} (default), the function will automatically select the appropriate
driver using \code{episerver_serverdetails("driver")}. See \code{\link[odbc]{odbcListDrivers}}
to identify which drivers you have access to.}

\item{max_attempts}{Integer to specify the number of attempts which will be made
to connect to the server. Defaults to \code{episerver_connect} default.  Workaround
for concurrency bug in RStudio.}

\item{...}{
  Arguments passed on to \code{\link{episerver_connect}}
  \describe{
    \item{\code{encrypt}}{Character string (\code{"Yes"} / \code{"No"}) or \code{NULL}. Sets the
ODBC \code{Encrypt} keyword, and takes effect only for ODBC Driver 18 or later.
When \code{NULL} (default) the value is resolved, in order, from
\code{getOption("actepir.encrypt")}, then the \code{ACTEPIR_ENCRYPT} environment
variable, then \code{"Yes"}. Set the option or the environment variable to change
the default for every EpiServer connection on a machine or in a session without
passing this argument to each function (see Details).}
    \item{\code{trust_certificate}}{Character string (\code{"Yes"} / \code{"No"}) or \code{NULL}.
Sets the ODBC \code{TrustServerCertificate} keyword, for ODBC Driver 18 or later.
When \code{NULL} (default) the value is resolved from
\code{getOption("actepir.trust_certificate")}, then \code{ACTEPIR_TRUST_CERTIFICATE},
then \code{"Yes"}. EpiServer presents a self-signed certificate, so this must be
\code{"Yes"} whenever \code{encrypt = "Yes"} on Driver 18.}
  }}
}
\description{
This function is a wraparound function which combines the \code{episerver_getlabels} and \code{dplyr::collect} functions and feeds them to the \code{applydatalabels} function.  It will allow the user to collect the result of a lazy query with associated labels when the query is placed against an EpiServer location.  The function will always search EpiServer and cannot be used for other servers of file locations.
}
\details{
Almost all non-administrative users will not need to use the lbl_* arguments.  In most cases, when accessing the "Analysis" database, you will not need to supply the "dataset" argument.
}
\examples{
\dontrun{
# One-step process: apply labels as part of collect() 
# statement when a dbplyr lazy object is passed to it
dat = episerver_connect("APC") \%>\% 
  collect_withlabels()

# One-step process: apply labels as part of collect() 
# statement but source labels from another location
dat = episerver_connect("APC") \%>\% 
  collect_withlabels(
    dataset = "APC", 
    lbl_table = "DataLabelsNew", 
    lbl_schema = "dbo", 
    lbl_db = "HospitalAPC"
    )
}
  
}
\keyword{episerver}
