% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dictionary.R
\name{create_dictionary}
\alias{create_dictionary}
\title{Create a data dictionary}
\usage{
create_dictionary(x, vars = NULL)
}
\arguments{
\item{x}{The dataframe of data containing labels. Argument should be supplied as an object.}

\item{vars}{The column in the dataframe that you wish to limit the extraction to. Argument should be supplied as an object.}
}
\description{
This function allow the users to extract variable and value labels into a flattened dataframe structure.
}
\details{
It assumes that the labels conform to the haven_labelled class and uses the \code{sjlabelled::get_label} and \code{sjlabelled::get_labels} functions.  This is a convenience wraparound function.
}
\examples{
df <- tibble::tibble(
  gender = haven::labelled(c(1, 2, 1), c(Male = 1, Female = 2),
                           label = "Gender of the respondent"),
  status = haven::labelled(c(1, 2, 2), c(Single = 1, Married = 2),
                           label = "Marital status")
  )

# Get flattened dictionary from already-labelled dataset
dictionary = create_dictionary(df)

# Limit dictionary to a single variable name
dictionary = create_dictionary(df,"gender")

# Limit dictionary to a list of variable names
dictionary = create_dictionary(df,c("gender","status"))
}
\keyword{labels}
