% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acthd_cols.R
\name{.acthd_cols}
\alias{.acthd_cols}
\title{Function to extract ACTHD colors as hex codes}
\usage{
.acthd_cols(...)
}
\arguments{
\item{...}{Character names of acthd_colors}
}
\description{
Function to extract ACTHD colors as hex codes
}
\keyword{internal}
