% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i_acthd_pal.R
\name{.acthd_pal}
\alias{.acthd_pal}
\title{Return function to interpolate ACTHD color palette}
\usage{
.acthd_pal(palette = "blue_turq", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in acthd_palettes}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\description{
Return function to interpolate ACTHD color palette
}
\keyword{internal}
