% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/episerver_browse.R
\name{episerver_browse_stop}
\alias{episerver_browse_stop}
\title{Stop the Background EpiServer Browser}
\usage{
episerver_browse_stop()
}
\value{
Invisibly returns \code{TRUE} if a process was stopped and
\code{FALSE} otherwise.
}
\description{
Stops the background process started by \code{\link{episerver_browse}}.
Has no effect if no background browser is running.
}
\examples{
\dontrun{
episerver_browse_stop()
}

}
\seealso{
\code{\link{episerver_browse}}
}
\author{
Warren Holroyd
}
\keyword{episerver}
\keyword{interactive}
