% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/episerver_dsr.R
\name{episerver_dsr_stop}
\alias{episerver_dsr_stop}
\title{Stop the Background DSR Calculator}
\usage{
episerver_dsr_stop()
}
\value{
Invisibly returns \code{TRUE} if a process was stopped and \code{FALSE}
otherwise.
}
\description{
Stops the background process started by \code{\link[=episerver_dsr]{episerver_dsr()}}. Has no effect if
no background calculator is running.
}
\examples{
\dontrun{
episerver_dsr_stop()
}

}
\seealso{
\code{\link[=episerver_dsr]{episerver_dsr()}}
}
\author{
Warren Holroyd
}
\keyword{episerver}
\keyword{interactive}
