% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/episerver_lazytable.R
\name{episerver_lazytable}
\alias{episerver_lazytable}
\title{Alias an EpiServer table as a \code{dbplyr} lazy table}
\usage{
episerver_lazytable(con, table, schema = "dbo", db = "Analysis")
}
\arguments{
\item{con}{The connection object which has defined the server connection.}

\item{table}{The name of the table to be aliased. Argument should be supplied as character.}

\item{schema}{The name of the schema to be aliased. Argument should be supplied as character. If not supplied, will default to \code{dbo}.}

\item{db}{The name of the database to be aliased. Argument should be supplied as character. If not supplied, will default to \code{Analysis}.}
}
\description{
This function allows the user to alias a specific table (with a \code{dbplyr} lazy query).  Note that this function is hard-coded to reference the EpiServer and cannot be used to contact a different server.  It will default to the \code{Analysis.dbo} namespace unless \code{schema} and \code{db} arguments are also supplied.
}
\examples{
\dontrun{
# Alias the APC table in default `Analysis` database
apc_tbl = episerver_lazytable("APC")

# Alias the `TestTable` table in non-default `AnalysisArchive` database
test_tbl = episerver_lazytable("TestTable", db="AnalysisArchive")
}

}
\keyword{episerver}
