% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/force_update.R
\name{force_update}
\alias{force_update}
\title{Force an update to actepir}
\usage{
force_update()
}
\description{
This allows the user to force a reinstallation of the actepir local package
}
\keyword{force_update}
