% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxy_actgov_off.R
\name{proxy_actgov_off}
\alias{proxy_actgov_off}
\title{Switch off ACTGOV proxy}
\usage{
proxy_actgov_off()
}
\description{
Useful for quickly nullifying Sys.setenv() proxy variables.
}
\examples{
proxy_actgov_off()

}
\keyword{proxy}
