% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxy_actgov_toggle.R
\name{proxy_actgov_toggle}
\alias{proxy_actgov_toggle}
\title{Toggle ACTGOV proxy}
\usage{
proxy_actgov_toggle(toggle, proxyaddress = NULL)
}
\arguments{
\item{toggle}{The instruction to switch on or off.  Argument should be supplied as character. Valid values are 'on' or 'off'.}

\item{proxyaddress}{The proxy address.  Argument should be supplied as character.  If left null, a known default will be applied.  Not required for toggling off.}
}
\description{
Useful for quickly setting Sys.setenv() proxy variables to the ACTGOV proxy server.
}
\examples{
proxy_actgov_toggle("on")

}
\keyword{proxy}
