% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acthd_ggplot_autotheme.R
\name{acthd_ggplot_autotheme}
\alias{acthd_ggplot_autotheme}
\title{Function to auto format ggplot object}
\usage{
acthd_ggplot_autotheme(x)
}
\description{
Function to auto format ggplot object
}
