% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acthd_gt_theme.R
\name{acthd_gt_theme}
\alias{acthd_gt_theme}
\title{Color theme for gt tables using ACT Health Directorate palettes}
\usage{
acthd_gt_theme(x, palette = "web_blues", reverse = FALSE)
}
\arguments{
\item{x}{The \code{gt} object to be themed}

\item{palette}{Character name of ACT palette. See 'details' for valid palette names.}

\item{reverse}{Boolean indicating whether the palette should be reversed.}
}
\description{
This function allows the user to apply pre-defined theme to a st object.
}
\details{
Palettes are intentionally designed from dark to light (meaning the darker colour will fill the table header, with the exception of \code{spectral}).  Applying the \code{reverse} argument may not lead to desirable results as tables can look odd with light headers and dark row lines.
}
\examples{
# Normal theme application
my_gt_table |>
  acthd_gt_theme()
  
# Reversed application
my_gt_table |>
  acthd_gt_theme(palette = "web_purples", reverse = TRUE)
 
}
\keyword{acthd}
\keyword{gt}
\keyword{table}
