% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acthd_list_palettes.R
\name{acthd_list_palettes}
\alias{acthd_list_palettes}
\title{Get ACTHD palette details}
\usage{
acthd_list_palettes(detail = "names", objtype = "list")
}
\arguments{
\item{detail}{What level of detail to be returned.  Valid options are \code{names} or \code{all}. Default is \code{names}.}

\item{objtype}{The type of object to be returned.  Valid options are \code{list} or \code{df}.  Default is \code{list}.}
}
\description{
This function will allow the user to get palette details.
}
\examples{
# Two-step process: download labels first 
# and save to object, then apply labels as 
# last step in downloading data
lab = episerver_getlabels("APC")
dat = episerver_connect("APC") \%>\% 
  collect() \%>\% 
  applydatalabels(lab)

}
\keyword{palettes}
