% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applydatalabels.R
\name{applydatalabels}
\alias{applydatalabels}
\title{Bulk apply labels to data}
\usage{
applydatalabels(data = parent.frame(), labels = NULL)
}
\arguments{
\item{data}{The dataframe of data which the labels will be applied to. Argument should be supplied as an object.}

\item{labels}{The dataframe containing the labels to be applied. Argument should be supplied as an object.}
}
\description{
This function will allow the user to apply value and variable labels in bulk to a dataset.
}
\details{
It assumes that the labelset is formatted according to those stored on EpiServer.  Columns should be:

"VarName" : the variable name
"VarType" : indicating whether the row specifies a variable or response option (value) label. an  Valid values in this column are "var" (for variable label) and "opt" (for value label)
"DataCode" : containing the underlying code (level)
"LabelName" : the variable or value label

All columns should be character.  VarNames must be in the same case as the column names in the data they are to be applied to.
}
\examples{
# Two-step process: download labels first and save to 
# object, then apply labels as last step in downloading data
lab = episerver_getlabels("APC")
dat = episerver_connect("APC") \%>\% 
  collect() \%>\% 
  applydatalabels(lab)

}
\keyword{labels}
