% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autofactor.R
\name{autofactor}
\alias{autofactor}
\title{Automatic factor conversion}
\usage{
autofactor(
  dfdata = parent.frame(),
  dflabels,
  vars,
  order = NULL,
  suffix = NULL,
  nolabel = NULL
)
}
\arguments{
\item{dfdata}{DATAFRAME. The dataframe containing the column(s) requiring factor conversion}

\item{dflabels}{DATAFRAME. The dataframe containing the data dictionary/labels}

\item{vars}{STRING VECTOR. A vector containing the column names that should be factor converted}

\item{order}{TRUE/FALSE. Should the resulting factor levels be ordered?}

\item{suffix}{STRING. Suffix to the resulting factor column name to distinguish it from the original. Not supplying will overwrite the original column.}

\item{nolabel}{STRING. The name of the level which unmatched levels are assigned to.}
}
\description{
This function will allow the user to convert a column to a factor and automatically set levels based on a supplied data dictionary. 
The dictionary must conform to the DataLabels structure used on EpiServer.
}
\examples{
varlist <- c("var1","var2")
mydata <- mydata \%>\% 
  autofactor(
    dflabels = mylabelsdf,
    vars = varlist,
    order = TRUE,
    suffix = "_X",
    nolabel = "UL"
    )
}
\keyword{autofactor}
