% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_withlabels.R
\name{collect_withlabels}
\alias{collect_withlabels}
\title{Download labels from EpiServer for a given dataset}
\usage{
collect_withlabels(dataset, lbl_db = NULL, lbl_schema = NULL, lbl_table = NULL)
}
\arguments{
\item{dataset}{The name of the dataset that labels are associated with. Argument should be supplied as character or object.}

\item{lbl_db}{The name of the database to be connected to. Argument should be supplied as character. If not supplied, will default to `Analysis`.}

\item{lbl_schema}{The name of the schema to be connected to. Argument should be supplied as character. If not supplied, will default to `ref`.}

\item{lbl_table}{The name of the table where the labels exist. Argument should be supplied as character. If not supplied, will default to `DataLabels`.}
}
\description{
This function will allow the user to download the labels file for a given dataset.  The function will always search EpiServer and cannot be used for other servers of file locations.
}
\details{
Almost all non-administrative users will not need to use the lbl_* arguments.
}
\examples{
# One-step process: apply labels as part of collect() 
# statement when a dbplyr lazy object is passed to it
dat = episerver_connect("APC") \%>\% 
  collect_withlabels()

# One-step process: apply labels as part of collect() 
# statement but source labels from another location
dat = episerver_connect("APC") \%>\% 
  collect_withlabels(
    dataset = "APC", 
    lbl_table = "DataLabelsNew", 
    lbl_schema = "dbo", 
    lbl_db = "HospitalAPC"
    )
  
}
\keyword{episerver}
