% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/episerver_getlabels.R
\name{episerver_getlabels}
\alias{episerver_getlabels}
\title{Download labels from EpiServer for a given dataset}
\usage{
episerver_getlabels(
  dataset,
  lbl_table = NULL,
  lbl_schema = NULL,
  lbl_db = NULL,
  driver = NULL,
  max_attempts = NULL,
  ...
)
}
\arguments{
\item{dataset}{The name of the dataset that labels are associated with. Argument
should be supplied as character or object.  Mandatory.}

\item{lbl_table}{The name of the table where the labels exist. Argument should be
supplied as character. If not supplied, will default to \code{DataLabels}.}

\item{lbl_schema}{The name of the schema to be connected to. Argument should be
supplied as character. If not supplied, will default to \code{ref}.}

\item{lbl_db}{The name of the database to be connected to. Argument should be
supplied as character. If not supplied, will default to \code{Analysis}.}

\item{driver}{Character string or \code{NULL}. ODBC driver to use for the connection.
If \code{NULL} (default), the function will automatically select the appropriate
driver using \code{episerver_serverdetails("driver")}. See \code{\link[odbc]{odbcListDrivers}}
to identify which drivers you have access to.}

\item{max_attempts}{Integer to specify the number of attempts which will be made
to connect to the server. Defaults to \code{episerver_connect} default.  Workaround
for concurrency bug in RStudio.}

\item{...}{
  Arguments passed on to \code{\link{episerver_connect}}
  \describe{
    \item{\code{encrypt}}{Character string (\code{"Yes"} / \code{"No"}) or \code{NULL}. Sets the
ODBC \code{Encrypt} keyword, and takes effect only for ODBC Driver 18 or later.
When \code{NULL} (default) the value is resolved, in order, from
\code{getOption("actepir.encrypt")}, then the \code{ACTEPIR_ENCRYPT} environment
variable, then \code{"Yes"}. Set the option or the environment variable to change
the default for every EpiServer connection on a machine or in a session without
passing this argument to each function (see Details).}
    \item{\code{trust_certificate}}{Character string (\code{"Yes"} / \code{"No"}) or \code{NULL}.
Sets the ODBC \code{TrustServerCertificate} keyword, for ODBC Driver 18 or later.
When \code{NULL} (default) the value is resolved from
\code{getOption("actepir.trust_certificate")}, then \code{ACTEPIR_TRUST_CERTIFICATE},
then \code{"Yes"}. EpiServer presents a self-signed certificate, so this must be
\code{"Yes"} whenever \code{encrypt = "Yes"} on Driver 18.}
  }}
}
\description{
This function will allow the user to download the labels file for a given dataset.  The function will always search EpiServer and cannot be used for other servers of file locations.
}
\details{
Almost all non-administrative users will not need to use the lbl_* arguments.
}
\examples{
# Import APC labels from default `Analysis` database
labels = episerver_getlabels("APC")

# Import APC labels from non-default location
labels = episerver_getlabels(
  dataset = "APC", 
  lbl_table = "DataLabelsNew", 
  lbl_schema = "dbo", 
  lbl_db = "HospitalAPC"
  )
 
}
\keyword{episerver}
