% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{episerver_serverdetails}
\alias{episerver_serverdetails}
\title{INTERNAL FUNCTIONS}
\usage{
episerver_serverdetails(feature)
}
\description{
Get EpiServer details
This function  allows the user quickly locate details about the EpiServer, such as address and port.
The feature/detail to be returned. Can be one of: "driver", "server", "port", "default_db" or "default_schema". Argument should be supplied as character.
episerver_serverdetails("port")
}
